requirements.txt
pony
python_telegram_bot
pytz
requests

Pillow